Put the parsing config into the adapter and put the adapter client into a separate file so we don't pollute the index manager functions
